requirements.txt
Flask==3.0.3
Flask-Cors==5.0.0
python-dotenv==1.0.1
gradio_client==1.3.0
groq==0.11.0
gunicorn==23.0.0
requests==2.32.3
urllib3==2.2.3
certifi==2024.8.30
charset-normalizer==3.3.2
idna==3.10
Werkzeug==3.0.4
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5